requirements.txt
django==4.1.7
djangorestframework==3.14.0
mysqlclient==2.1.1
Pillow==9.5.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.1.0